+++ example/lib/minimal.dart
@@ -5,17 +5,15 @@ class MinimalistApp extends StatelessWidget {
   MinimalistApp({Key? key}) : super(key: key);
 
   late final router = UrlRouter(
-    onGeneratePages: (router) {
-      return [
-        MaterialPage(child: Center(child: Text(router.url))),
-      ];
-    },
+    onGeneratePages: (router) => [
+      MaterialPage(child: Center(child: Text(router.url))),
+    ],
   );
 
   @override
   Widget build(BuildContext context) {
     return MaterialApp.router(
-      routeInformationParser: UrlRouter.parser,
+      routeInformationParser: UrlRouteParser(),
       routerDelegate: router,
     );
   }