+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## 0.2.0
+- changed scaffoldBuilder tp builder
+- made onGeneratePages optional, added assert
+
 ## 0.1.1
 - Improve queryParams support when performing pop/push
 - Add queryParams setter